Fix test failure when default gems are installed into separate directory.
However, this test does not appear to have much of value, because
previously, it tested that the default gem executable is overriden by
the user installed package. This of course cannot happen in case where
default gems are stored into independent directory structure.

It could be different, if both the default as well as the user installed
gem installed their executable into common directory, e.g. `Gem.bindir`.